Update user serialization and add days_of_week property***
This commit updates the `to_representation` method in the `TeacherField` class to include the `user` object's `id` and `name` attributes. It also updates the `to_internal_value` method to use `user_id` instead of `id` when querying the `Teacher` model.

Additionally, this commit adds a new property `days_of_week` to the `SubjectPeriod` model, which returns the related `SubjectPeriodWeekday` objects.